requirements.txt
<<<<<<< HEAD
# Ubuntu (Linux) 전용
# 설치: pip install -r requirements.txt

=======
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
absl-py==2.3.1
astunparse==1.6.3
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
gast==0.4.0
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.74.0
h5py==3.15.1
idna==3.11
joblib==1.5.3
<<<<<<< HEAD
keras==2.15.0
=======
keras==3.12.1
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
mediapipe==0.10.9
<<<<<<< HEAD
ml-dtypes==0.2.0
=======
ml-dtypes==0.3.2
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
MouseInfo==0.1.3
namex==0.1.0
numpy==1.24.3
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
optree==0.18.0
<<<<<<< HEAD
packaging>=24.0
=======
packaging
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
pillow==12.1.0
protobuf==3.20.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
PyGetWindow==0.0.9
Pygments==2.19.2
PyMsgBox==2.0.1
pynput==1.8.1
pyparsing==3.3.2
pyperclip==1.11.0
PyQt6==6.10.2
PyQt6-Qt6==6.10.1
PyQt6_sip==13.11.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.1
rsa==4.9.1
<<<<<<< HEAD
=======
rubicon-objc==0.5.3
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sounddevice==0.5.5
<<<<<<< HEAD
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
=======
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-estimator==2.13.0
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
<<<<<<< HEAD
wrapt==1.14.2
=======
wrapt==2.0.1
>>>>>>> d1bd67f5dcb6706aacd57c6cdd4a254dd5041311
